ref(linter): move each lint into own file
Move each lint into their own file along with some changes to the snap
locations.

Next step is to move the tests for each lint into their own file.
